+++ customersupport/templates/index.html
@@ -7,6 +7,7 @@
 
 <body>
 <h1>Hello World</h1>
+<a href='/session'>Click me!</a>
 </body>
 
 </html>
