+++ customersupport/views.py
@@ -6,8 +6,13 @@ from flask import json
 
 from customersupport import app
 from customersupport.database import db_session
+from customersupport.session_controller import SessionController
 # from customersupport.models import CallLog
 
 @app.route('/')
 def index():
     return render_template('index.html')
+
+@app.route('/session')
+def sessionCall():
+    return json.dumps(SessionController.getEmployeeStubbed().json())
